More awesome code by aarti

+++ AKSwiftSlideMenu-master/AKSwiftSlideMenu/Discount.swift
@@ -13,14 +13,12 @@ class Discount: NSObject
     // MARK: Properties
     
     var shopName: String = ""
-    
     var shopAddress: String = ""
     var discount: String = ""
     var discountBanner: String = ""
-    var lat: Int = 0
-    var long: Int = 0
-    var likes: Int = 0
-    
+    var lat: Double = 0
+    var long: Double = 0
+    var likes: Int16 = 0
     // MARK: Initialization
 //    
 //    init?(shopName: String, shopAddress: String, discount: String ,discountBanner: String, lat: Int, long: Int, likes: Int)
